refactor: reorganize test folder and extract tutorial constants
